fix(exportaudit): split canvas_draw.go under 800-line gate, silence gosec on codemod IO
The sweep's keep markers pushed gui/canvas_draw.go to 802 lines,
tripping the large-files gate; the exported draw types moved to
canvas_draw_types.go. The tool's own ReadFile/WriteFile calls are
repo-relative paths it derived itself, not user input — marked
#nosec G304/G306 per repo convention.
